requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
drf-yasg===1.21.5